updated jest config to split to group of "Unit Tests" and "Integration Tests" and set up test environments accordingly
