+++ test/unit/ReactCodepen.test.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import {when} from 'jest-when'
 import {shallow} from "enzyme";
 
-import ReactCodepen from "../src";
+import ReactCodepen from "../../src";
 
 describe('<ReactCodepen/>', function () {
     it('should mount without error', function () {
